+++ src/api/java/com/nisovin/shopkeepers/api/shopkeeper/player/PlayerShopCreationData.java
@@ -9,6 +9,7 @@ import org.bukkit.entity.Player;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopCreationData;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopType;
 import com.nisovin.shopkeepers.api.shopobjects.ShopObjectType;
+import com.nisovin.shopkeepers.api.shopobjects.virtual.VirtualShopObjectType;
 
 /**
  * Shop creation data specific for player shops.
@@ -26,6 +27,10 @@ public class PlayerShopCreationData extends ShopCreationData {
 										Location spawnLocation, BlockFace targetedBlockFace, Block chest) {
 		super(creator, shopType, objectType, spawnLocation, targetedBlockFace);
 		Validate.isTrue(shopType instanceof PlayerShopType, "Shop type has to be a PlayerShopType!");
+		// chest needs to be located in a world, which is only available for non-virtual shops:
+		// TODO decouple shopkeeper/shop object location from chest location? (allows chest in different world, and
+		// virtual player shopkeepers connected to a chest located in a world)
+		Validate.isTrue(!(objectType instanceof VirtualShopObjectType), "Cannot create virtual player shops!");
 		Validate.notNull(chest, "Chest is null!");
 		Validate.isTrue(spawnLocation.getWorld().equals(chest.getWorld()),
 				"Chest is located in a different world than the spawn location!");