+++ src/api/java/com/nisovin/shopkeepers/api/shopobjects/virtual/VirtualShopObject.java
@@ -1,0 +1,9 @@
+package com.nisovin.shopkeepers.api.shopobjects.virtual;
+
+import com.nisovin.shopkeepers.api.shopobjects.ShopObject;
+
+/**
+ * A {@link ShopObject} that is not present in any world.
+ */
+public interface VirtualShopObject extends ShopObject {
+}